Use the docker entrypoint from official mysql img
This entry point script does some neat functionality. The main reason
for adding this is that, if you don't have a db setup in the volume
/var/lib/mysql there is no way for creating a database. This is because
by default the user created by mysql is not allowed to log in from
another host that is not localhost. This makes imposible for another
container to create databeses.